# Better elaboration of pattern-matchings on primitive projections