<!-- Thank you for your contribution.
     Make sure you read the contributing guide and fill this template. -->

Better representation of cases on primitive projections. Work in progress.
Now using let-ins consistently for the thing to destruct and each projection
to avoid blowups. Currently based on 8.7 for testing on current developments.

<!-- Keep what applies -->
**Kind:** feature / performance.

<!-- If this is a bug fix, make sure the bug was reported beforehand. -->
Fixes / closes #5701


<!-- If this is a feature pull request / breaks compatibility: -->
<!-- (Otherwise, remove these lines.) -->
- [x] Corresponding documentation was added / updated.
- [x] Entry added in CHANGES.