@mattam82 2018-03-08
Actually it happens quite a bunch in HoTT put I doubt they test "real" computations as well