@mattam82 2018-03-08
Let's run a pendulum on HoTT to see if this changes anything.  https://ci.inria.fr/coq/view/benchmarking/job/benchmark-part-of-the-branch/367

I think there were real world examples in some discussion about this, maybe by @silene ?