@silene 2018-03-08
> maybe by @silene ?

Since I am not a user of primitive projections, I doubt so. It does raise a question though, since It is not clear to me by reading the commit message or the code. How many times is the destructed expression evaluated? I hope the answer is just one.