@mattam82 2018-03-08
Found it, it was @robbertkrebbers in #6723 which should be closed by this as well.