@ppedrot 2018-05-05
Benchmark:
```
┌──────────────────────────┬─────────────────────────┬───────────────────────────────────────┬───────────────────────────────────────┬─────────────────────────┬─────────────────┐
│                          │      user time [s]      │              CPU cycles               │           CPU instructions            │  max resident mem [KB]  │   mem faults    │
│                          │                         │                                       │                                       │                         │                 │
│             package_name │     NEW     OLD PDIFF   │            NEW            OLD PDIFF   │            NEW            OLD PDIFF   │     NEW     OLD PDIFF   │ NEW OLD PDIFF   │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-sf-vfa │   26.23   26.73 -1.87 % │    72452181430    72583945815 -0.18 % │    93751792711    93969428314 -0.23 % │  541708  541372 +0.06 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-sf-lf │   20.68   20.98 -1.43 % │    56500106258    56508917491 -0.02 % │    73527922113    73750702926 -0.30 % │  429028  429036 -0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│             coq-compcert │  871.32  876.40 -0.58 % │  2416659818349  2428187305992 -0.47 % │  3482080827764  3482931984248 -0.02 % │ 1333148 1333116 +0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│           coq-coquelicot │   83.65   83.86 -0.25 % │   231455705359   230946907498 +0.22 % │   275386754606   275630364289 -0.09 % │  719700  719364 +0.05 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│            coq-fiat-core │  111.72  111.95 -0.21 % │   311539736414   310912777013 +0.20 % │   379648707446   379713600313 -0.02 % │  499892  499956 -0.01 % │   4   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-flocq │  263.05  263.31 -0.10 % │   729508328085   730651875294 -0.16 % │   903395449178   903474363791 -0.01 % │ 1245684 1245504 +0.01 % │  15   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                  coq-vst │ 3851.32 3850.33 +0.03 % │ 10715521091012 10712391369507 +0.03 % │ 13998668256120 14002553512341 -0.03 % │ 2232788 2234764 -0.09 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-fingroup │   65.40   65.38 +0.03 % │   179820872455   179183038062 +0.36 % │   239472367274   239755258070 -0.12 % │  589080  588872 +0.04 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│         coq-fiat-parsers │  705.79  705.26 +0.08 % │  1953963190624  1952864731313 +0.06 % │  2980178485070  2980628868665 -0.02 % │ 3376152 3376156 -0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│              coq-bignums │   80.11   80.01 +0.12 % │   220646648171   220479277714 +0.08 % │   285975701220   286346486046 -0.13 % │  528652  528572 +0.02 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│       coq-mathcomp-field │  461.34  460.64 +0.15 % │  1280699595314  1278769361298 +0.15 % │  2064439557958  2064736142751 -0.01 % │  814400  814388 +0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-geocoq │ 3102.58 3094.99 +0.25 % │  8624712201399  8604717501530 +0.23 % │ 13814873950740 13820300793051 -0.04 % │ 1239720 1239684 +0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-ssreflect │   45.46   45.34 +0.26 % │   123764195195   123644982610 +0.10 % │   145135344470   145292837021 -0.11 % │  537212  537276 -0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│         coq-math-classes │  229.76  229.14 +0.27 % │   631737561390   629062630819 +0.43 % │   786100416114   786292825885 -0.02 % │  533292  533228 +0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-solvable │  203.03  202.40 +0.31 % │   560811859316   558743559971 +0.37 % │   770519043831   770743815681 -0.03 % │  866176  857172 +1.05 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│          coq-fiat-crypto │ 4128.44 4112.53 +0.39 % │ 11453959442089 11408846803236 +0.40 % │ 18417446861639 18396535720140 +0.11 % │ 3179028 3189260 -0.32 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-color │  725.44  722.63 +0.39 % │  2008901065997  2003396246824 +0.27 % │  2384238733283  2384159936541 +0.00 % │ 1446604 1446500 +0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│     coq-mathcomp-algebra │  186.50  185.73 +0.41 % │   516268366544   513587893606 +0.52 % │   692919943666   693061466168 -0.02 % │  645200  644952 +0.04 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│ coq-mathcomp-real-closed │  172.77  171.99 +0.45 % │   478730134544   476629357630 +0.44 % │   698349982361   698614230363 -0.04 % │  837280  837196 +0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│              coq-unimath │ 1295.88 1289.65 +0.48 % │  3591541746122  3569065804873 +0.63 % │  6041325359360  6040227293295 +0.02 % │ 1072004 1071984 +0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-character │  276.52  274.88 +0.60 % │   766633628554   762015710826 +0.61 % │  1087222349665  1087409406052 -0.02 % │ 1121416 1121608 -0.02 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│      coq-formal-topology │   37.88   37.62 +0.69 % │   101451468081   101500562311 -0.05 % │   126203806292   126374676046 -0.14 % │  482480  482364 +0.02 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-odd-order │ 1440.14 1426.50 +0.96 % │  4000590541771  3962564315297 +0.96 % │  6663248500969  6663419240974 -0.00 % │ 1382056 1382024 +0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                 coq-hott │  335.00  331.60 +1.03 % │   931264505392   923379439731 +0.85 % │  1426449052154  1426563678944 -0.01 % │  588964  589040 -0.01 % │   1   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                 coq-corn │ 1588.00 1568.65 +1.23 % │  4392794923148  4342615576810 +1.16 % │  6449669251085  6450539132474 -0.01 % │  874260  874180 +0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-sf-plf │   54.33   53.62 +1.32 % │   149067907675   148595923943 +0.32 % │   187601180745   187749868483 -0.08 % │  519820  519948 -0.02 % │   0   0  +nan % │
└──────────────────────────┴─────────────────────────┴───────────────────────────────────────┴───────────────────────────────────────┴─────────────────────────┴─────────────────┘


PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = 34837e89c25f035c80ad3f6ef1ab4f9c29af9d1a
OLD = 892fe9fc259bb812d10dfb5af40dec02412ff45e
```
I'm slightly surprised by the overhead in CoRN, I'll rerun a bench on that one.